[ES-1303159] Create a new http client for each different connection (#581)
It is possible for multiple connections running in different threads to unintentionally share the same HTTP client. When one connection closes its resources—including the shared HTTP client—it can leave the other connection with a closed or invalid HTTP client. This causes issues when the other connection attempts to use the client, as it no longer functions properly due to being closed by the first connection.

- create a new client for each connection
- added a concurrency test suite