[ES-1436915] Pass context to public classes and their dependencies (#806)
For public classes and interfaces, we can't make assumptions about the threading environment in which they will be used. Specifically, host applications may utilize these classes, such as ResultSet, in a thread that is different from the one that created the JDBC connection resource.

In the current implementation of certain APIs, along with the execution flow of DatabricksResultSet, there is an expectation that the thread will automatically be populated with an internal context object. This assumption is problematic, given the scenario described above, where threads may not have the expected context object set by default.

To address this issue, this pull request explicitly passes the context object to the areas where it was previously missing, correcting the faulty implementation.

Key changes in this PR:

- The metadata result set builder class has been refactored from a static utility class to an instance-based class, with the context object now stored as a member variable.

- The context object is now passed explicitly to the DatabricksResultSet class, ensuring that it functions correctly in multithreaded environments.

These changes eliminate the previous assumption about the thread's context