Fix flaky test ChunkLinkDownloadServiceTest#testBatchDownloadChaining (#811)
The link download futures were being completed because the link download chain was triggered eagerly. When getLinkForChunk is called, the link download future is evaluated. If the future is already complete, it checks the status of the chunk. The test was not mocking the chunk, which caused a NullPointerException (NPE). Locally, the test passed because the link download futures weren't completing before getLinkForChunk was called. However, in the GitHub runs, the futures were completing, which led to a thread synchronization issue.